readme update

+++ services/fileCleanupService.js
@@ -8,6 +8,10 @@ export const cleanUploads = (uploadDir) => {
                 console.error("Errore durante la lettura della cartella uploads:", err);
                 return next(err);
             }
+            if (files.length === 0) {
+                return next();
+            }
+            // Elimina i file nella cartella
             files.forEach(file => {
                 const filePath = path.join(uploadDir, file);
                 fs.unlink(filePath, err => {
